Added all functional

+++ notes/views.py
@@ -1,10 +1,11 @@
+from django.http import HttpResponse
 from django.shortcuts import render, redirect
 from .forms import LoginForm, RegisterForm
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login
 from django.contrib.auth.decorators import login_required
 from .models import Note
-
+from django.template.loader import render_to_string
 
 # Create your views here.
 def index(request):
@@ -26,6 +27,8 @@ def register(request):
 
 
 def login_user(request):
+    if request.user.is_authenticated:
+        return redirect('notes')
     if request.method == 'POST':
         username = request.POST['username']
         password = request.POST['password']
@@ -34,43 +37,55 @@ def login_user(request):
         user = User.objects.filter(username=username, password=password).first()
         if user is not None:
             login(request, user)
-            return redirect('/allNotes')
+            return redirect('/notes')
 
     form = LoginForm()
     return render(request, 'login.html', {'form': form})
 
 
 @login_required
-def add_note(request):
+def note(request):
+    if request.method == 'GET':
+        note_id = int(request.GET.get('id', 0))
+        if note_id > 0:
+            selected_note = Note.objects.filter(id=note_id).first()
+            if not selected_note or selected_note.user_id != request.user:
+                return HttpResponse(render_to_string('error.html', {'error_message': 'Unauthorized!'}), status=401)
+            return render(request, 'note.html', { 'note': selected_note })
+        return render(request, 'note.html')
+    
     if request.method == 'POST':
-        note_id = int(request.POST.get('id', 0))
-        user_id = request.user.id
+        id = request.POST.get('id', 0)
+        note_id = int(0 if not id else id)
         title = request.POST.get('title')
         content = request.POST.get('content', '')
 
         if note_id > 0:
-            document = Note.objects.get(pk=note_id)
-            document.title = title
-            document.content = content
-            document.save()
+            note = Note.objects.get(pk=note_id)
+            note.title = title
+            note.content = content
+            note.save()
 
-            return redirect('/?id=%i' % note_id)
+            return redirect('/note/?id=%i' % note_id)
         else:
-            document = Note.objects.create(title=title, content=content, user_id=user_id)
-
-            return redirect('/?id=%i' % document.id)
-    return render(request, 'addNote.html')
+            note = Note.objects.create(title=title, content=content, user_id=request.user)
+            return redirect('/note/?id=%i' % note.id)
+    return render(request, 'notes.html')
 
 
 @login_required
-def view_notes(request):
+def notes(request):
     notes = Note.objects.filter(user_id=request.user.id)
-    return render(request, 'allNotes.html', notes)
+    context = {
+        'notes': notes,
+    }
+    return render(request, 'notes.html', context)
 
 
 @login_required
-def remove_note(request):
-    note_id = request.POST.get('remove')
-    note = Note.objects.get(id=note_id)
+def remove_note(request, id):
+    note = Note.objects.filter(id=id)
+    if not note or note.user_id != request.user:
+        return HttpResponse(render_to_string('error.html', {'error_message': 'Unauthorized!'}), status=401)
     note.delete()
-    return redirect('/allNotes')
+    return redirect('/notes')
